@@ -1,73 +1,68 @@
 ﻿using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reactive;
 using System.Threading.Tasks;
+using System.Windows.Navigation;
+using Tauron.Features;
 
 namespace Akka.MGIHelper.Core.FanControl.Bus
 {
-    public sealed class MessageBus : IAsyncDisposable
+    public sealed class MessageBus : IDisposable
     {
-        private readonly ConcurrentDictionary<Type, List<object>> _handlers = new ConcurrentDictionary<Type, List<object>>();
-        private readonly object _lock = new object();
+        private readonly Dictionary<Type, List<object>> _handlers = new();
+        private readonly object _lock = new();
 
-        public async ValueTask DisposeAsync()
+        public void Dispose()
         {
-            foreach (var dis in _handlers.SelectMany(l => l.Value))
-                switch (dis)
-                {
-                    case IAsyncDisposable asyncDisposable:
-                        await asyncDisposable.DisposeAsync();
-                        break;
-                    case IDisposable disposable:
-                        disposable.Dispose();
-                        break;
-                }
+            foreach (var dis in _handlers.SelectMany(l => l.Value).OfType<IDisposable>())
+                dis.Dispose();
 
             _handlers.Clear();
         }
 
-        public IAsyncDisposable Subscribe<TMsg>(IHandler<TMsg> handler)
+        public void Subscribe<TMsg>(IHandler<TMsg> handler)
         {
             var msgType = typeof(TMsg);
-            lock (_lock)
-            {
-                if (_handlers.TryGetValue(msgType, out var list))
-                    list.Add(handler);
-                else
-                    _handlers.TryAdd(msgType, new List<object> {handler});
-            }
 
-            return new SubscribeDispose<TMsg>(handler, _handlers[msgType], _lock);
+            if (_handlers.TryGetValue(msgType, out var list))
+                list.Add(handler);
+            else
+                _handlers.TryAdd(msgType, new List<object> {handler});
         }
 
-        public async Task Publish<TMsg>(TMsg msg)
+        //public async Task<Unit> Publish<TMsg>(TMsg msg)
+        //{
+        //        foreach (var handler in _handlers[typeof(TMsg)].OfType<IHandler<TMsg>>()) await handler.Handle(msg, this).ConfigureAwait(false);
+        //        return Unit.Default;
+        //}
+
+        public async Task<TState> Publish<TState, TMsg>(StatePair<TMsg, TState> msg)
         {
-            foreach (var handler in _handlers[typeof(TMsg)].OfType<IHandler<TMsg>>()) await handler.Handle(msg, this).ConfigureAwait(false);
+            foreach (var handler in _handlers[typeof(TMsg)].OfType<IHandler<TMsg>>()) await handler.Handle(msg.State, this).ConfigureAwait(false);
+            return msg.State;
         }
 
-        private sealed class SubscribeDispose<TMsg> : IAsyncDisposable
-        {
-            private readonly IHandler<TMsg> _handler;
-            private readonly List<object> _handlers;
-            private readonly object _locker;
+        //private sealed class SubscribeDispose<TMsg> : IDisposable
+        //{
+        //    private readonly IHandler<TMsg> _handler;
+        //    private readonly List<object> _handlers;
+        //    private readonly object _locker;
 
-            public SubscribeDispose(IHandler<TMsg> handler, List<object> handlers, object locker)
-            {
-                _handler = handler;
-                _handlers = handlers;
-                _locker = locker;
-            }
+        //    public SubscribeDispose(IHandler<TMsg> handler, List<object> handlers, object locker)
+        //    {
+        //        _handler = handler;
+        //        _handlers = handlers;
+        //        _locker = locker;
+        //    }
 
-            public ValueTask DisposeAsync()
-            {
-                lock (_locker)
-                {
-                    _handlers.Remove(_handler);
-                }
-
-                return new ValueTask(Task.CompletedTask);
-            }
-        }
+        //    public void Dispose()
+        //    {
+        //        lock (_locker)
+        //        {
+        //            _handlers.Remove(_handler);
+        //        }
+        //    }
+        //}
     }
 }
