@@ -1,12 +1,4 @@
 ﻿namespace Akka.MGIHelper.Core.FanControl.Events
 {
-    public class ClockEvent
-    {
-        public ClockEvent(ClockState clockState)
-        {
-            ClockState = clockState;
-        }
-
-        public ClockState ClockState { get; }
-    }
+    public sealed record ClockEvent(ClockState ClockState);
 }
