@@ -1,6 +1,5 @@
 ﻿namespace Akka.MGIHelper.Core.FanControl.Events
 {
-    public class TickEvent
-    {
-    }
+    public sealed record TickEvent;
+
 }
