@@ -1,18 +1,27 @@
 ﻿using System;
-using System.Linq.Expressions;
+using System.Reactive.Linq;
 using System.Threading.Tasks;
 using Akka.Actor;
 using Akka.MGIHelper.Core.Configuration;
 using Akka.MGIHelper.Core.FanControl.Bus;
 using Akka.MGIHelper.Core.FanControl.Components;
 using Akka.MGIHelper.Core.FanControl.Events;
+using Tauron;
+using Tauron.Features;
 
 namespace Akka.MGIHelper.Core.FanControl
 {
-    public class FanControl : ReceiveActor
+    public sealed class FanControl : ActorFeatureBase<FanControlOptions>
     {
-        public FanControl(FanControlOptions options)
+        public static IPreparedFeature New(FanControlOptions options)
+            => Feature.Create(() => new FanControl(), options);
+
+        private FanControl()
+        { }
+
+        protected override void Config()
         {
+            var options = CurrentState;
             var parent = Context.Parent;
 
             var messageBus = new MessageBus();
@@ -22,10 +31,10 @@ namespace Akka.MGIHelper.Core.FanControl
             messageBus.Subscribe(new DataFetchComponent(options));
 
             messageBus.Subscribe(new TrackingEventDeliveryComponent(e =>
-            {
-                parent.Tell(e);
-                return Task.CompletedTask;
-            }));
+                                                                    {
+                                                                        parent.Tell(e);
+                                                                        return Task.CompletedTask;
+                                                                    }));
 
             messageBus.Subscribe(new PowerComponent());
             messageBus.Subscribe(new CoolDownComponent());
@@ -34,18 +43,14 @@ namespace Akka.MGIHelper.Core.FanControl
             messageBus.Subscribe(new StandByCoolingComponent(options));
 
             messageBus.Subscribe(new FanControlComponent(options, e =>
-            {
-                parent.Tell(new FanStatusChange(e));
-                return Task.CompletedTask;
-            }));
+                                                                  {
+                                                                      parent.Tell(new FanStatusChange(e));
+                                                                      return Task.CompletedTask;
+                                                                  }));
 
-            Receive<ClockEvent>(async msg => await messageBus.Publish(msg));
-        }
+            Stop.Subscribe(_ => messageBus.Dispose());
 
-        private void CreateComponent<TComp>(Expression<Func<TComp>> fac)
-            where TComp : ActorBase
-        {
-            Context.ActorOf(Props.Create(fac), typeof(TComp).Name); //.WithDispatcher("akka.CallingThread")
+            Receive<ClockEvent>(obs => obs.SelectMany(async p => await messageBus.Publish(p)));
         }
     }
 }
