@@ -1,61 +1,48 @@
-﻿using System.Diagnostics;
+﻿using System;
+using System.Diagnostics;
+using System.Reactive.Linq;
 using System.Threading;
 using Akka.Actor;
-using Akka.Event;
+using Tauron;
+using Tauron.Features;
 
 namespace Akka.MGIHelper.Core.ProcessManager
 {
-    public sealed class TrackedProcessActor : ReceiveActor
+    public sealed class TrackedProcessActor : ActorFeatureBase<TrackedProcessActor.TrackedProcessState>
     {
-        private readonly Timer _exitCheck;
-        private readonly int _id;
+        public sealed record TrackedProcessState(Timer ExitCheck, int Id, string ProcessName, Process Target);
 
-        private readonly ILoggingAdapter _log = Context.GetLogger();
-        private readonly string _processName;
-        private readonly Process _target;
+        public static IPreparedFeature New(Process toTrack, int id, string name)
+            => Feature.Create(() => new TrackedProcessActor(), c => new TrackedProcessState(new Timer(_ => c.Self.Tell(InternalCheckProcess.Inst), null, 3000, 1000), id, name, toTrack));
 
-        public TrackedProcessActor(Process process)
+        private TrackedProcessActor() { }
+
+        protected override  void Config()
         {
-            _target = process;
+            Stop.Subscribe(_ =>
+                           {
+                               CurrentState.ExitCheck.Dispose();
+                               CurrentState.Target.Dispose();
+                           });
+            
+            Receive<InternalCheckProcess>(obs => obs.Where(e => e.State.Target.HasExited)
+                                                    .Select(_ => InternalProcessExit.Inst)
+                                                    .ToSelf());
 
-            _processName = _target.ProcessName;
-            _id = process.Id;
-            _log.Info("Track Process {Name} Startet: {Id}", _processName, _id);
-
-            var self = Self;
-            _exitCheck = new Timer(o => self.Tell(new InternalCheckProcess()), null, 1000, 1000);
-
-            Receive<InternalProcessExit>(ProcessOnExited);
-            Receive<InternalCheckProcess>(CheckProcess);
+            Receive<InternalProcessExit>(obs => obs.Do(p => Log.Info("Track Process {Name} Exited: {Id}", p.State.ProcessName, p.State.Id))
+                                                   .Do(_ => Context.Stop(Self))
+                                                   .Select(p => new ProcessExitMessage(p.State.Target, p.State.ProcessName, p.State.Id))
+                                                   .ToParent());
         }
 
-        private void CheckProcess(InternalCheckProcess obj)
+        private sealed record InternalProcessExit
         {
-            if (_target.HasExited)
-                Self.Tell(new InternalProcessExit());
+            public static readonly InternalProcessExit Inst = new();
         }
 
-        private void ProcessOnExited(InternalProcessExit msg)
+        private sealed record InternalCheckProcess
         {
-            _log.Info("Track Process {Name} Exited: {Id}", _processName, _id);
-
-            Context.Parent.Tell(new ProcessExitMessage(_target, _processName, _id));
-            Context.Stop(Self);
-        }
-
-
-        protected override void PostStop()
-        {
-            _exitCheck.Dispose();
-            _target.Dispose();
-        }
-
-        private sealed class InternalProcessExit
-        {
-        }
-
-        private sealed class InternalCheckProcess
-        {
+            public static readonly InternalCheckProcess Inst = new();
         }
     }
 }
